Materials: rendermaterials.py reorganization vs. renderutils.py
Some code moving between the two files, to prepare some more major
evolutions into rendermaterials
